add basic cron job
