+++ backend/django/daily_database_pull.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+dt=$(date '+%d/%m/%Y %H:%M:%S');
+echo "$dt"
+echo "HIIIIIIII"
+#python manage.py ingest_database_data 4
